Modified to enable calculation of divalent ions

+++ src/Common/CompMs.Common.Lipidomics.SourceGenerator/Sentence.cs
@@ -1,6 +1,6 @@
 ﻿namespace CompMs.Common.Lipidomics.SourceGenerator;
 
 internal sealed class Sentence {
-    public (Term, int)[] Terms { get; set; } = [];
+    public (Term, double)[] Terms { get; set; } = [];
     public string Raw { get; set; } = string.Empty;
 }
